+++ examples/IRremoteExtensionTest/IRremoteExtensionClass.cpp
@@ -36,11 +36,11 @@
  * Otherwise you may get warnings like "type 'struct IRData' itself violates the C++ One Definition Rule"
  */
 #if !defined(RAW_BUFFER_LENGTH)
-// For air condition remotes it requires 600 (maximum for 2k RAM) to 750. Default is 112 if DECODE_MAGIQUEST is enabled, otherwise 100.
+// For air condition remotes it requires 750. Default is 200.
 #  if (defined(RAMEND) && RAMEND <= 0x4FF) || (defined(RAMSIZE) && RAMSIZE < 0x4FF)
-#define RAW_BUFFER_LENGTH  180 // Requires 360 bytes RAM
+#define RAW_BUFFER_LENGTH  360
 #  elif (defined(RAMEND) && RAMEND <= 0x8FF) || (defined(RAMSIZE) && RAMSIZE < 0x8FF)
-#define RAW_BUFFER_LENGTH  600 // Requires 1200 bytes RAM
+#define RAW_BUFFER_LENGTH  750
 #  endif
 #endif