+++ src/IRremoteInt.h
@@ -64,20 +64,21 @@
 #define USE_DEFAULT_FEEDBACK_LED_PIN    0
 
 /**
- * The length of the buffer where the IR timing data is stored before decoding
- * 100 is sufficient for most standard protocols, but air conditioners often send a longer protocol data stream
+ * The RAW_BUFFER_LENGTH determines the length of the byte buffer where the received IR timing data is stored before decoding.
+ * 100 is sufficient for standard protocols up to 48 bits, with 1 bit consisting of one mark and space plus 1 byte for initial gap, 2 bytes for header and 1 byte for stop bit.
+ * 48 bit protocols are PANASONIC, KASEIKYO, SAMSUNG48, RC6.
+ * 32 bit protocols like NEC, SAMSUNG, WHYNTER, SONY(20), LG(28) requires a buffer length of 68.
+ * 16 bit protocols like BOSEWAVE, DENON, FAST, JVC, LEGO_PF, RC5, SONY(12 or 15) requires a buffer length of 36.
+ * MAGIQUEST requires a buffer length of 112.
+ * Air conditioners often send a longer protocol data stream up to 750 bits.
  */
 #if !defined(RAW_BUFFER_LENGTH)
 #  if (defined(RAMEND) && RAMEND <= 0x8FF) || (defined(RAMSIZE) && RAMSIZE < 0x8FF)
 // for RAMsize <= 2k
-#    if defined(DECODE_MAGIQUEST)
-#define RAW_BUFFER_LENGTH  112  // MagiQuest requires 224 bytes.
-#    else
-#define RAW_BUFFER_LENGTH  100  ///< Length of raw duration buffer. 100 -> 200 bytes. Must be even. 100 supports up to 48 bit codings inclusive 1 start and 1 stop bit.
-#    endif
+#define RAW_BUFFER_LENGTH  200  ///< Length of raw duration buffer. Must be even. 100 supports up to 48 bit codings inclusive 1 start and 1 stop bit.
 #  else
 // For undefined or bigger RAMsize
-#define RAW_BUFFER_LENGTH  750 // Requires 1500 bytes RAM. The value for air condition remotes.
+#define RAW_BUFFER_LENGTH  750 // The value for air condition remotes.
 #  endif
 #endif
 #if RAW_BUFFER_LENGTH % 2 == 1
@@ -89,6 +90,29 @@ typedef uint_fast8_t IRRawlenType;
 #else
 typedef unsigned int IRRawlenType;
 #endif
+
+/*
+ * Use 8 bit buffer for IR timing in 50 ticks units.
+ * It is save to use 8 bit if RECORD_GAP_TICKS < 256, since any value greater 255 is interpreted as frame gap of 12750 us.
+ * The default for frame gap is currently 8000!
+ * But if we assume that for most protocols the frame gap is way greater than the biggest mark or space duration,
+ * we can choose to use a 8 bit buffer even for frame gaps up to 200000 us.
+ * This enables the use of 8 bit buffer even for more some protocols like B&O or LG air conditioner etc.
+ */
+#if RECORD_GAP_TICKS <= 400 // Corresponds to RECORD_GAP_MICROS of 200000. A value of 255 is foolproof, but we assume, that the frame gap is
+typedef uint8_t IRRawbufType; // all timings up to the gap fit into 8 bit.
+#else
+typedef uint16_t IRRawbufType; // The gap does not fit into 8 bit ticks value. This must not be a reason to use 16 bit for buffer, but it is at least save.
+#endif
+
+#if (__INT_WIDTH__ < 32)
+typedef uint32_t IRRawDataType;
+#define BITS_IN_RAW_DATA_TYPE   32
+#else
+typedef uint64_t IRRawDataType;
+#define BITS_IN_RAW_DATA_TYPE   64
+#endif
+
 /****************************************************
  * Declarations for the receiver Interrupt Service Routine
  ****************************************************/
@@ -103,7 +127,7 @@ typedef unsigned int IRRawlenType;
  * Only StateForISR needs to be volatile. All the other fields are not written by ISR after data available and before start/resume.
  */
 struct irparams_struct {
-    // The fields are ordered to reduce memory over caused by struct-padding
+    // The fields are ordered to reduce memory overflow caused by struct-padding
     volatile uint8_t StateForISR;       ///< State Machine state
     uint_fast8_t IRReceivePin;          ///< Pin connected to IR data from detector
 #if defined(__AVR__)
@@ -115,17 +139,11 @@ struct irparams_struct {
     void (*ReceiveCompleteCallbackFunction)(void); ///< The function to call if a protocol message has arrived, i.e. StateForISR changed to IR_REC_STATE_STOP
 #endif
     bool OverflowFlag;                  ///< Raw buffer OverflowFlag occurred
-    IRRawlenType rawlen;               ///< counter of entries in rawbuf
-    uint16_t rawbuf[RAW_BUFFER_LENGTH]; ///< raw data / tick counts per mark/space, first entry is the length of the gap between previous and current command
+    IRRawlenType rawlen;                ///< counter of entries in rawbuf
+    uint16_t initialGapTicks;   ///< Tick counts of the length of the gap between previous and current IR frame. Pre 4.4: rawbuf[0].
+    IRRawbufType rawbuf[RAW_BUFFER_LENGTH]; ///< raw data / tick counts per mark/space. With 8 bit we can only store up to 12.7 ms. First entry is empty to be backwards compatible.
 };
 
-#if (__INT_WIDTH__ < 32)
-typedef uint32_t IRRawDataType;
-#define BITS_IN_RAW_DATA_TYPE   32
-#else
-typedef uint64_t IRRawDataType;
-#define BITS_IN_RAW_DATA_TYPE   64
-#endif
 #include "IRProtocol.h"
 
 /*
@@ -171,7 +189,7 @@ struct decode_results {
     bool isRepeat;              // deprecated, moved to decodedIRData.flags ///< True if repeat of value is detected
 
 // next 3 values are copies of irparams_struct values - see above
-    uint16_t *rawbuf;       // deprecated, moved to decodedIRData.rawDataPtr->rawbuf ///< Raw intervals in 50uS ticks
+    uint16_t *rawbuf;           // deprecated, moved to decodedIRData.rawDataPtr->rawbuf ///< Raw intervals in 50uS ticks
     uint_fast8_t rawlen;        // deprecated, moved to decodedIRData.rawDataPtr->rawlen ///< Number of records in rawbuf
     bool overflow;              // deprecated, moved to decodedIRData.flags ///< true if IR raw code too long
 };