+++ changelog.md
@@ -3,6 +3,8 @@ The latest version may not be released!
 See also the commit log at github: https://github.com/Arduino-IRremote/Arduino-IRremote/commits/master
 
 # 4.4.0
+- Using 8 bit raw timing buffer for all timings except frame gap (former rawbuf[0]).
+- Renamed decodedIRData.initialGap to decodedIRData.initialGapTicks.
 - sendNEC() and sendNEC2() now accepts 16 bit command to better map to NECext protocol found in IRDB databases.
 - ir_DistanceWidthProtocol() now decodes up to 10 ms mark or spaces if RAM is bigger than 2 k.
 - Improved sensitivity and decoding of PULSE_DISTANCE + PULSE_WIDTH protocols.